requirements.txt
pre-commit
black
GitPython